Fix warnings; add macros SUPPRESS_WARNINGS_START and SUPPRESS_WARNINGS_END
Add missing SPDX headers and LICENSES/Zlib. Remove unused variables and
add a cast in GLTFBuilder::makeTextureParameters() to fix a warning in a
formerly signed vs. unsigned integer comparison. Modernize old-style
("K&R") function definitions. Use the new macros SUPPRESS_WARNINGS_START
and SUPPRESS_WARNINGS_END to silence a self-move warning in
shared_ptr_test.cpp (don't suppress the self-move as it is precisely
what is being tested).

+++ simgear/scene/model/ReaderWriterGLTF.cxx
@@ -1,7 +1,6 @@
-/*
- * SPDX-FileCopyrightText: Copyright (C) 2021 - 2024 Fernando García Liñán
- * SPDX-License-Identifier: LGPL-2.0-or-later
- */
+
+// SPDX-License-Identifier: LGPL-2.0-or-later
+// SPDX-FileCopyrightText: 2021-2024 Fernando García Liñán
 
 #ifdef HAVE_CONFIG_H
 #  include <simgear_config.h>
@@ -330,7 +329,7 @@ struct GLTFBuilder {
         makeChild(texNode, "type")->setValue("2d");
         makeChild(texNode, "image")->setStringValue(absFileName);
 
-        if (texture.sampler >= 0 && texture.sampler < model.samplers.size()) {
+        if ((texture.sampler >= 0) && (static_cast<size_t>(texture.sampler) < model.samplers.size())) {
             const tinygltf::Sampler& sampler = model.samplers[texture.sampler];
 
             // Build a single sampler property
